backend/requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file='requirements\base.txt' 'requirements\base.in'
#
amqp==5.3.1
    # via kombu
asgiref==3.11.1
    # via
    #   channels
    #   channels-redis
    #   django
    #   django-cors-headers
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.4
    # via celery
celery==5.6.2
    # via
    #   -r requirements/base.in
    #   django-celery-beat
    #   django-celery-results
channels==4.3.2
    # via
    #   -r requirements/base.in
    #   channels-redis
channels-redis==4.3.0
    # via -r requirements/base.in
click==8.3.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
cron-descriptor==2.0.6
    # via django-celery-beat
dj-database-url==3.1.0
    # via -r requirements/base.in
django==5.2.11
    # via
    #   -r requirements/base.in
    #   channels
    #   dj-database-url
    #   django-celery-beat
    #   django-celery-results
    #   django-cors-headers
    #   django-filter
    #   django-tenants
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-celery-beat==2.8.1
    # via -r requirements/base.in
django-celery-results==2.6.0
    # via -r requirements/base.in
django-cors-headers==4.9.0
    # via -r requirements/base.in
django-environ==0.12.0
    # via -r requirements/base.in
django-filter==25.2
    # via -r requirements/base.in
django-tenants==3.10.0
    # via -r requirements/base.in
django-timezone-field==7.2.1
    # via django-celery-beat
djangorestframework==3.16.1
    # via
    #   -r requirements/base.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements/base.in
drf-spectacular==0.29.0
    # via -r requirements/base.in
hiredis==3.3.0
    # via -r requirements/base.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
kombu==5.6.2
    # via celery
msgpack==1.1.2
    # via channels-redis
packaging==26.0
    # via kombu
pillow==12.1.1
    # via -r requirements/base.in
reportlab==4.4.10
    # via -r requirements/base.in (PDF generation for quotes)
prompt-toolkit==3.0.52
    # via click-repl
psycopg[binary]==3.3.2
    # via -r requirements/base.in
psycopg-binary==3.3.2
    # via psycopg
pyjwt==2.11.0
    # via djangorestframework-simplejwt
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via celery
pyyaml==6.0.3
    # via drf-spectacular
redis==7.1.1
    # via
    #   -r requirements/base.in
    #   channels-redis
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via cron-descriptor
tzdata==2025.3
    # via
    #   django
    #   django-celery-beat
    #   kombu
    #   psycopg
    #   tzlocal
tzlocal==5.3.1
    # via celery
uritemplate==4.2.0
    # via drf-spectacular
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.6.0
    # via prompt-toolkit
whitenoise==6.11.0
    # via -r requirements/base.in
